QEXDEV-332 - Changes made after review from Ina and Nikita
