--- playwright-workshop/src/test/java/com/devexperts/in/qatesting/HomePage.java
@@ -1,49 +1,0 @@
-package com.devexperts.in.qatesting;
-
-import com.microsoft.playwright.Locator;
-import com.microsoft.playwright.Page;
-
-import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-public class HomePage {
-
-    Page page;
-    private static final String HOMEPAGE_TITLE_SELECTOR = ".header-title-content";
-
-    private static final String BALANCE_SELECTOR = "#balanceInfoNumber";
-    private static final String TOTAL_FUNDS_SELECTOR = "#balanceTotalNumber";
-
-    public HomePage(Page page){
-        this.page = page;
-    }
-
-    private Locator getHeaderHomepage(){
-        return page.locator(HOMEPAGE_TITLE_SELECTOR);
-    }
-
-    public void assertHeaderHomepage(String pageHeader){
-        assertAll("Home Page Checks",
-                () ->assertThat(getHeaderHomepage()).hasText(pageHeader),
-                () ->assertThat(getHeaderHomepage()).isVisible());
-    }
-
-    // Balance Section
-
-    private Locator getBalance(){
-        return page.locator(BALANCE_SELECTOR);
-    }
-
-    private Locator getTotalFunds(){
-        return page.locator(TOTAL_FUNDS_SELECTOR);
-    }
-
-    public void assertInitialBalances(String deposit){
-        assertAll("Initial Balances are equal to selected Deposit",
-                () ->assertThat(getBalance()).hasText(deposit),
-                () ->assertThat(getBalance()).isVisible(),
-                () ->assertThat(getTotalFunds()).hasText(deposit),
-                () ->assertThat(getTotalFunds()).isVisible());
-    }
-
-}
