+++ playwright-workshop/src/test/java/com/devexperts/in/qatesting/tests/LoginTest.java
@@ -1,0 +1,83 @@
+package com.devexperts.in.qatesting.tests;
+
+import com.devexperts.in.qatesting.configuration.PropertiesProvider;
+import com.devexperts.in.qatesting.pages.HomePage;
+import com.devexperts.in.qatesting.pages.LoginPage;
+import com.microsoft.playwright.Browser;
+import com.microsoft.playwright.Page;
+import com.microsoft.playwright.Playwright;
+import org.junit.jupiter.api.*;
+
+import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
+
+public class LoginTest {
+
+    private static final String USERNAME_DATA = PropertiesProvider.getProperty("test.user");
+    private static final String PASSWORD_DATA = PropertiesProvider.getProperty("test.password");
+    private static final String USERNAME_WRONG_DATA = PropertiesProvider.getProperty("wrong.test.user");
+    private static final String PASSWORD_WRONG_DATA = PropertiesProvider.getProperty("wrong.test.password");
+
+    private static Playwright playwright;
+    private Browser browser;
+    private Page page;
+
+    @BeforeAll
+    public static void createPlaywright(){
+        playwright = Playwright.create();
+    }
+
+    @BeforeEach
+    public void setUp(){
+        browser = playwright.chromium().launch();
+        page = browser.newPage();
+        page.navigate(PropertiesProvider.getProperty("base.url"));
+    }
+
+    @Test
+    public void testSuccessfulLogin(){
+
+        LoginPage loginPage = new LoginPage(page);
+        loginPage.login(USERNAME_DATA, PASSWORD_DATA);
+
+        HomePage homePage = new HomePage(page);
+        assertThat(homePage.getHeaderHomepage()).hasText("Home Test Task");
+    }
+
+    @Test
+    public void testLoginWithoutUserData(){
+
+        LoginPage loginPage = new LoginPage(page);
+        loginPage.login("", "");
+
+        assertThat(loginPage.getLoginStatus()).hasText("Please enter valid credentials:");
+    }
+
+    @Test
+    public void testLoginWithWrongUser(){
+
+        LoginPage loginPage = new LoginPage(page);
+        loginPage.login(USERNAME_WRONG_DATA, PASSWORD_WRONG_DATA);
+
+        assertThat(loginPage.getLoginStatus()).hasText("Wrong user! User " + USERNAME_WRONG_DATA + " not found.");
+    }
+
+    @Test
+    public void testLoginWithWrongPassword(){
+
+        LoginPage loginPage = new LoginPage(page);
+        loginPage.login(USERNAME_DATA, PASSWORD_WRONG_DATA);
+
+        assertThat(loginPage.getLoginStatus()).hasText("Wrong password! Correct password is: " + PASSWORD_DATA);
+    }
+
+    @AfterEach
+    public void tearDown(){
+        page.close();
+        browser.close();
+    }
+
+    @AfterAll
+    public static void closePlaywright(){
+        playwright.close();
+    }
+}
