+++ playwright-workshop/src/test/resources/config.properties
@@ -10,7 +10,12 @@ slow.mo=50
 base.url=https://qa-testing.in.devexperts.com/internship/
 
 # Username for the test user.
-test.user=changeit
+test.user=gmshn.edu@gmail.com
+wrong.test.user=wrong_username
 
 # Password for the test user.
-test.password=changeit
+test.password=autoQAedu@2024
+wrong.test.password=wrong_password
+
+# Deposit for the test user.
+test.deposit=5000.00
